refactor(redis-instance): rename redis_instance to redis-instance
in general, using hyphens over underscores in filenames are recommend
